outsourced Select component to UI
