+++ src/pages/Settings/ProfileSettings/ProfileSettings.module.css
@@ -74,9 +74,7 @@
 }
 
 .inputTextarea:hover,
-.inputTextarea:focus,
-.displayInterests:hover,
-.displayInterests:focus {
+.inputTextarea:focus {
     outline: 0;
     border: 2px solid #6831de;
 }
@@ -92,13 +90,6 @@
     flex-direction: column;
     align-items: center;
     margin: 0 auto;
-}
-
-.displayInterests {
-    border: 2px solid grey;
-}
-.displayInterests div:first-of-type {
-    border: transparent;
 }
 
 @media (max-width: 700px) {
